api/ml-api/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
flask-restx==1.3.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
nltk==3.9.1
numpy==2.1.1
packaging==24.1
python-dotenv==1.0.1
pytz==2024.2
rake-nltk==1.0.6
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
threadpoolctl==3.5.0
tqdm==4.66.5
urllib3==2.3.0
Werkzeug==3.0.4
